backend/requirements.txt
async-timeout==4.0.3
blinker==1.8.2
cachelib==0.13.0
cachetools==4.2.4
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Session==0.8.0
google-api-core==1.16.0
google-api-python-client==1.8.4
google-auth==1.12.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.65.0
gspread==5.7.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgspec==0.18.6
oauthlib==3.2.2
protobuf==5.28.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dotenv==1.0.0
pytz==2024.2
redis==5.1.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.0
six==1.16.0
uritemplate==3.0.1
urllib3==2.2.3
Werkzeug==3.0.4
